+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# gh development version
+# gh 1.3.0
 
 * gh now shows the correct number of records in its progress bar when
   paginating (#147).